backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Base de datos
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Validación y configuración
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilidades
python-dotenv==1.0.1
